+++ website/src/lib/prismadb.ts
@@ -3,7 +3,7 @@ declare global {
   // eslint-disable-next-line no-var
   var prisma: PrismaClient | undefined;
 }
-
+console.trace()
 const client = globalThis.prisma || new PrismaClient();
 if (process.env.NODE_ENV !== "production") {
   globalThis.prisma = client;
