+++ website/src/pages/api/leaderboard.ts
@@ -1,5 +1,5 @@
 import { withoutRole } from "src/lib/auth";
-import { createApiClient } from "src/lib/oasst_api_client";
+import { createApiClient } from "src/lib/oasst_client_factory";
 import { LeaderboardTimeFrame } from "src/types/Leaderboard";
 
 /**
