+++ website/src/pages/api/valid_labels.ts
@@ -1,5 +1,5 @@
 import { withoutRole } from "src/lib/auth";
-import { createApiClient } from "src/lib/oasst_api_client";
+import { createApiClient } from "src/lib/oasst_client_factory";
 
 /**
  * Returns the set of valid labels that can be applied to messages.
